fix: minor fixes

+++ src/controllers/post/delete-post-by-id.ts
@@ -4,5 +4,11 @@ export async function deletePostById(params: {
   postRepo: IPostRepo;
   postId: string;
 }) {
+  const post = await params.postRepo.getPostById(params.postId);
+
+  if (!post) {
+    throw new Error('Post not found');
+  }
+
   await params.postRepo.deletePostById(params.postId);
 }
